requirements.txt
pandas
numpy
scikit-learn
matplotlib
pymysql
flask
prometheus_client
paho-mqtt
pyserial